renamed IAM and removed smells
