+++ Apps/Common/src/Common.csproj
@@ -1,4 +1,4 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>    
     <TargetFramework>netcoreapp3.1</TargetFramework>
@@ -58,7 +58,6 @@
     <Folder Include="Constants\" />
     <Folder Include="Services\" />
     <Folder Include="Jobs\" />
-    <Folder Include="Delegates\" />
     <Folder Include="Authorization\Admin\" />
   </ItemGroup>
 </Project>
