+++ Apps/Common/src/Models/AuthServer/UserRepresentation.cs
@@ -1,5 +1,5 @@
 //-------------------------------------------------------------------------
-// Copyright © 2019 Province of British Columbia
+// Copyright © 2020 Province of British Columbia
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -13,12 +13,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 //-------------------------------------------------------------------------
-namespace HealthGateway.Common.Models.IAM
+namespace HealthGateway.Common.Models.AuthServer
 {
     using System;
 
     /// <summary>
-    /// Class that represents the user model in the IAM account system.
+    /// Class that represents the user model in the Authorization Server's user accounts.
     /// </summary>
     public class UserRepresentation
     {
@@ -29,32 +29,32 @@ namespace HealthGateway.Common.Models.IAM
 
         /// <summary>
         /// Gets or sets the user's email.
-        /// </summary>      
+        /// </summary>
         public string? Email { get; set; }
 
         /// <summary>
         /// Gets or sets the user's first name.
-        /// </summary>          
+        /// </summary>
         public string? FirstName { get; set; }
 
         /// <summary>
         /// Gets or sets the user's last name.
-        /// </summary>  
+        /// </summary>
         public string? LastName { get; set; }
 
         /// <summary>
-        /// Gets or sets the user's roles in the IAM application realm.
-        /// </summary>       
-        public string?[] RealmRoles { get; set; } = new string[0];
+        /// Gets the user's roles in the AuthServer's application realm.
+        /// </summary>
+        public ArraySegment<string> RealmRoles { get; } = Array.Empty<string>();
 
         /// <summary>
         /// Gets or sets the user's username.
-        /// </summary> 
+        /// </summary>
         public string? Username { get; set; }
 
         /// <summary>
         /// Gets or sets the user's unique account identifier.
-        /// </summary> 
+        /// </summary>
         public string? UserId { get; set; }
     }
